+++ server/pkg/app_settings/events.go
@@ -3,9 +3,9 @@ package app_settings
 import (
 	"encoding/json"
 
-	"github.com/freshteapot/learnalist-api/server/pkg/dripfeed"
 	"github.com/freshteapot/learnalist-api/server/pkg/event"
 	"github.com/freshteapot/learnalist-api/server/pkg/openapi"
+	"github.com/freshteapot/learnalist-api/server/pkg/spaced_repetition/dripfeed"
 )
 
 func (s AppSettingsService) OnEvent(entry event.Eventlog) {
